+++ typescript/packages/well-log-viewer/src/components/PlotDialog.tsx
@@ -84,18 +84,29 @@ export function createScaleItems(): ReactNode[] {
 function createColorItems(): ReactNode[] {
     return _createItems(colorItems);
 }
+
 export function createBooleanItems(): ReactNode[] {
     return _createItems(booleanItems);
 }
 
 function createColorTableItems(colorTables: ColorTable[]): ReactNode[] {
     const nodes: ReactNode[] = [];
-    for (const colorTable of colorTables) {
-        if (colorTable.discrete)
-            // skip discrete color tables
-            continue;
-        nodes.push(<option key={colorTable.name}>{colorTable.name}</option>);
-    }
+    if (!colorTables) {
+        console.error(
+            "colorTables is missed or empty in createColorTableItems()"
+        );
+    } else
+        for (const colorTable of colorTables) {
+            if (colorTable.discrete) continue; // skip discrete color tables
+            if (!colorTable.name) {
+                console.log(
+                    "colorTable.name is empty in createColorTableItems()"
+                );
+            }
+            nodes.push(
+                <option key={colorTable.name}>{colorTable.name}</option>
+            );
+        }
     return nodes;
 }
 
@@ -205,7 +216,8 @@ export class PlotPropertiesDialog extends Component<Props, State> {
                   inverseColor: "",
 
                   // for 'gradientfill' plot
-                  colorTable: this.props.wellLogView.props.colorTables[0].name,
+                  colorTable:
+                      this.props.wellLogView.props.colorTables?.[0]?.name,
                   inverseColorTable: undefined,
                   colorScale: undefined,
                   inverseColorScale: undefined,