+++ typescript/packages/well-log-viewer/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.7.0](https://github.com/equinor/webviz-subsurface-components/compare/well-log-viewer@1.6.0...well-log-viewer@1.7.0) (2024-04-05)
+
+
+### Features
+
+* Customizable WellLog Viewer Layout ([#1363](https://github.com/equinor/webviz-subsurface-components/issues/1363)) ([dc956e8](https://github.com/equinor/webviz-subsurface-components/commit/dc956e82f84e04f1719b859829301d8cf6b3b8aa)), closes [#1702](https://github.com/equinor/webviz-subsurface-components/issues/1702) [#1239](https://github.com/equinor/webviz-subsurface-components/issues/1239)
+
+# [1.6.0](https://github.com/equinor/webviz-subsurface-components/compare/well-log-viewer@1.5.5...well-log-viewer@1.6.0) (2024-04-04)
+
+
+### Features
+
+* Allow camera xy scaling ([#2002](https://github.com/equinor/webviz-subsurface-components/issues/2002)) ([0635f5e](https://github.com/equinor/webviz-subsurface-components/commit/0635f5e1f268a99ab0755c66522ba83f056f1365))
+
 ## [1.5.5](https://github.com/equinor/webviz-subsurface-components/compare/well-log-viewer@1.5.4...well-log-viewer@1.5.5) (2024-03-29)